move listener/subscriber to own namespaces
